% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_data_from_canvas.R
\name{bootleg_rogerhub}
\alias{bootleg_rogerhub}
\title{Calculate grade needed on final}
\usage{
bootleg_rogerhub(grade_worth_table, final_worth, goal)
}
\arguments{
\item{grade_worth_table}{A grade book table from scrape_html_for_grades or a table formatted like its return value.}

\item{final_worth}{A numeric value that the final is worth.}

\item{goal}{A numeric value that the goal grade is for the class.}
}
\value{
A numeric value that shows what you have to get on your final to get a certain grade in the class.
}
\description{
Scrapes the grades from Canvas and puts into a column.
}
