% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_data_from_canvas.R
\name{calc_grade_per_category}
\alias{calc_grade_per_category}
\title{Calculate the overall grades per category}
\usage{
calc_grade_per_category(grades)
}
\arguments{
\item{grades}{A grade book table from scrape_html_for_grades or a table formatted like its return value.}
}
\value{
A table with calculated overall grade per category from grade book table
}
\description{
calculate the overall grades per category
}
