% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_data_from_canvas.R
\name{calc_worth_per_category}
\alias{calc_worth_per_category}
\title{Calculate weight/worth per category}
\usage{
calc_worth_per_category(grade_table, worth_vector)
}
\arguments{
\item{grade_table}{A grade book table from calc_grade_per_category}

\item{worth_vector}{A numeric vector where each value is 1-100 which the shown percentages of how much each category is worth (in alphabetical order).}
}
\value{
A one column dataframe that has all the grades from Canvas. Note that if the
grade book is not complete then there may be null/missing values in the column. I
recommend you fix those by hand to match whatever you're doing.
}
\description{
Calculate weight/worth per category out of 100.
}
