% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_canvas.R
\name{get_html_element}
\alias{get_html_element}
\title{Converts html into strings, helper function for all the functions with html_filename as input}
\usage{
get_html_element(html_filename, element)
}
\arguments{
\item{html_filename}{Path to HTML file of Canvas page, recommended to place HTML file in the
same folder as where this is being used. Can also take Canvas HTML link but may not work due to
password protected websites.}

\item{element}{HTML element to extract}
}
\value{
The HTML in string form
}
\description{
Converts html into strings, helper function for all the functions with html_filename as input
}
