% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_canvas.R
\name{scrape_asgn_grades}
\alias{scrape_asgn_grades}
\title{Scrapes the grades}
\usage{
scrape_asgn_grades(html_filename)
}
\arguments{
\item{html_filename}{Path to HTML file of Canvas page, recommended to place HTML file in the
same folder as where this is being used. Can also take Canvas HTML link but may not work due to
log-ins sometimes.}
}
\value{
A one column tibble that has all the grades from Canvas. Note that if the
grade book is not complete then there may be null/missing values in the column. I
recommend you fix those by hand to match whatever you're doing.
}
\description{
Scrapes the grades
}
