% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_grade_breakdown.R
\name{scrape_grade_breakdown}
\alias{scrape_grade_breakdown}
\title{Scrapes the grade breakdown table on the right header of the Canvas grade page}
\usage{
scrape_grade_breakdown(html_filename)
}
\arguments{
\item{html_filename}{Path to HTML file of Canvas page, recommended to place HTML file in the
same folder as where this is being used. Can also take Canvas HTML link but may not work due to
log-ins sometimes.}
}
\value{
A 2 by X size tibble with categories and percentages as columns.
}
\description{
Scrapes the grade breakdown table on the right header of the Canvas grade page
}
