% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_canvas.R
\name{scrape_html_for_grades}
\alias{scrape_html_for_grades}
\title{Scrapes the HTML page for gradebook table}
\usage{
scrape_html_for_grades(html_filename)
}
\arguments{
\item{html_filename}{Path to HTML file of Canvas page, recommended to place HTML file in the
same folder as where this is being used. Can also take Canvas HTML link but may not work due to
log-ins sometimes.}
}
\value{
A table with all the assignments and their grades.
}
\description{
Scrapes the HTML page for gradebook table
}
